mobile/android/tests/browser/chrome/test_accounts.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=917942
Migrated from Robocop: https://bugzilla.mozilla.org/show_bug.cgi?id=1184186
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 917942</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript;version=1.7">

  Components.utils.import("resource://gre/modules/Accounts.jsm");

  add_task(function* () {
    let syncExists = yield Accounts.syncAccountsExist();
    info("Sync account exists? " + syncExists + "\n");
    let firefoxExists = yield Accounts.firefoxAccountsExist();
    info("Firefox account exists? " + firefoxExists + "\n");
    let anyExists = yield Accounts.anySyncAccountsExist();
    info("Any accounts exist? " + anyExists + "\n");

    // Only one account should exist.
    ok(!syncExists || !firefoxExists, "at least one account does not exist");
    is(anyExists, firefoxExists || syncExists, "sync/firefox account existence consistent with any existence");

    // TODO: How can this be cleaned up?
    //info("Launching setup.\n");
    //Accounts.launchSetup();
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=917942">Mozilla Bug 917942</a>
<br>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1184186">Migrated from Robocop testAccounts</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
